+++ test/unit/mysql_rep_cmp/get_db_tbl.py
@@ -22,7 +22,6 @@ import mock
 # Local
 sys.path.append(os.getcwd())
 import mysql_rep_cmp                            # pylint:disable=E0401,C0413
-import lib.gen_libs as gen_libs             # pylint:disable=E0401,C0413,R0402
 import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
@@ -60,8 +59,9 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
-        test_pre_802
-        test_pre_80
+        test_remove_tbl
+        test_ignore_db_tbl2
+        test_ignore_db_tbl
         test_812
         test_81
         test_802
@@ -69,8 +69,6 @@ class UnitTest(unittest.TestCase):
         test_with_db_tbl2
         test_with_db_tbl
         test_with_system_db_only3
-        test_with_system_db_only2
-        test_with_system_db_only
         test_with_empty_db_list
         test_with_multiple_dbs
         test_with_single_db
@@ -100,56 +98,84 @@ class UnitTest(unittest.TestCase):
         self.tbl_list2 = ["t1", "t2"]
         self.tbl_dict = [{"TABLE_NAME": "t2"}]
         self.tbl_dict2 = [{"TABLE_NAME": "t1"}, {"TABLE_NAME": "t2"}]
-        self.tbl_dict56 = [{"table_name": "t1"}, {"table_name": "t2"}]
         self.all_tbls = {"db1": ["t2"]}
         self.all_tbls2 = {"db1": ["t2"], "db2": ["t1"]}
         self.ign_dbs = ["systemdb"]
+
+        self.tbl_dict6 = [
+            {"TABLE_NAME": "tbl1"}, {"TABLE_NAME": "tbl2"},
+            {"TABLE_NAME": "tbl3"}]
+        self.tbl_dict8 = [
+            {"TABLE_NAME": "tbl1"}, {"TABLE_NAME": "tbl2"}]
+        self.ign_db_tbl6 = {"db1": ["tbl1", "tbl3"]}
+        self.ign_db_tbl7 = {"db2": ["tbl1", "tbl3"]}
+        self.tbl_list6 = ["tbl1", "tbl2", "tbl3"]
+        self.results6 = {"db1": ["tbl2"]}
+        self.results7 = {"db1": ["tbl1", "tbl2", "tbl3"]}
+        self.results8 = {"db1": ["tbl1", "tbl2"]}
+
         self.results = {"db1": ["t2"]}
         self.results2 = {"db1": ["t2"], "db2": ["t1"]}
         self.results3 = {}
         self.results4 = {"db1": ["t1", "t2"]}
 
     @mock.patch("mysql_rep_cmp.mysql_libs.fetch_tbl_dict")
-    def test_pre_802(self, mock_fetch):
+    def test_remove_tbl(self, mock_fetch):
 
-        """Function:  test_pre_802
+        """Function:  test_remove_tbl
 
-        Description:  Test in pre-8.0 version.
+        Description:  Test with remove tables passed that are not in the
+            database.
 
         Arguments:
 
         """
 
-        self.server.version = (5, 6)
-
-        mock_fetch.return_value = self.tbl_dict56
+        mock_fetch.return_value = self.tbl_dict8
 
         self.assertEqual(
             mysql_rep_cmp.get_db_tbl(
                 self.server, self.db_list2, ign_dbs=self.ign_dbs,
-                tbls=self.tbl_list2),
-            self.results4)
+                tbls=self.tbl_list6, ign_db_tbl=self.ign_db_tbl7),
+            self.results8)
 
     @mock.patch("mysql_rep_cmp.mysql_libs.fetch_tbl_dict")
-    def test_pre_80(self, mock_fetch):
+    def test_ignore_db_tbl2(self, mock_fetch):
 
-        """Function:  test_pre_80
+        """Function:  test_ignore_db_tbl2
 
-        Description:  Test in pre-8.0 version.
+        Description:  Test with ignoring some tables, but not in same database.
 
         Arguments:
 
         """
 
-        self.server.version = (5, 6)
-
-        mock_fetch.return_value = self.tbl_dict56
+        mock_fetch.return_value = self.tbl_dict6
 
         self.assertEqual(
             mysql_rep_cmp.get_db_tbl(
                 self.server, self.db_list2, ign_dbs=self.ign_dbs,
-                tbls=self.tbl_list),
-            self.results)
+                tbls=self.tbl_list6, ign_db_tbl=self.ign_db_tbl7),
+            self.results7)
+
+    @mock.patch("mysql_rep_cmp.mysql_libs.fetch_tbl_dict")
+    def test_ignore_db_tbl(self, mock_fetch):
+
+        """Function:  test_ignore_db_tbl
+
+        Description:  Test with ignoring some tables.
+
+        Arguments:
+
+        """
+
+        mock_fetch.return_value = self.tbl_dict6
+
+        self.assertEqual(
+            mysql_rep_cmp.get_db_tbl(
+                self.server, self.db_list2, ign_dbs=self.ign_dbs,
+                tbls=self.tbl_list6, ign_db_tbl=self.ign_db_tbl6),
+            self.results6)
 
     @mock.patch("mysql_rep_cmp.mysql_libs.fetch_tbl_dict")
     def test_812(self, mock_fetch):
@@ -285,44 +311,6 @@ class UnitTest(unittest.TestCase):
                 self.server, self.db_list4, ign_dbs=self.ign_dbs),
             self.results)
 
-    @mock.patch("mysql_rep_cmp.mysql_libs.fetch_db_dict")
-    def test_with_system_db_only2(self, mock_fetch):
-
-        """Function:  test_with_system_db_only2
-
-        Description:  Test with empty database list.
-
-        Arguments:
-
-        """
-
-        mock_fetch.return_value = self.fetch_db3
-
-        with gen_libs.no_std_out():
-            self.assertEqual(
-                mysql_rep_cmp.get_db_tbl(
-                    self.server, self.db_list, ign_dbs=self.ign_dbs),
-                self.results3)
-
-    @mock.patch("mysql_rep_cmp.get_all_dbs_tbls")
-    def test_with_system_db_only(self, mock_all):
-
-        """Function:  test_with_system_db_only
-
-        Description:  Test with system only database passed.
-
-        Arguments:
-
-        """
-
-        mock_all.return_value = self.all_tbls
-
-        with gen_libs.no_std_out():
-            self.assertEqual(
-                mysql_rep_cmp.get_db_tbl(
-                    self.server, self.db_list3, ign_dbs=self.ign_dbs),
-                self.results3)
-
     @mock.patch("mysql_rep_cmp.get_all_dbs_tbls")
     @mock.patch("mysql_rep_cmp.mysql_libs.fetch_db_dict")
     def test_with_empty_db_list(self, mock_fetch, mock_all):
@@ -364,7 +352,7 @@ class UnitTest(unittest.TestCase):
 
         """Function:  test_with_single_db
 
-        Description:  Test with single database.
+        Description:  Test with single database and get all tables in database.
 
         Arguments: